Huge path changes
This allows for paths to be much more sophisticated. More work needs to be done still, but so far we have: input, optional, and loop "blocks" for the path. Input blocks are still not fully implemented, and the pathFor method does not correctly work as of now, but should in the near future.